Move Login/Logout control into own file User
